Comentei os códigos e retirei uma função não utilizada no src/infos/jogador_1.c

+++ include/infos/inimigo_1.h
@@ -3,15 +3,18 @@
 
 #include "actions/atq_def_inimigo_1.h"
 
+//Declara a struct para as informações do inimigo 1
 typedef struct {
     char nome[20];
     double vida;
     double resistência;
 } info_inimigo_1;
 
+//Função para inicializar as informações do jogador 1
 void func_info_inimigo_1(info_inimigo_1 *ptr_info_inimigo_1);
 
-double escolha_ataque_inimigo_1(atq_def_inimigo_1 *ptr_atq_def_inimigo_1);
+//Função para inicializar as ações do jogador 1 
+int escolha_ataque_inimigo_1(atq_def_inimigo_1 *ptr_atq_def_inimigo_1);
 
 #endif
